Revert "Change naming back to . instead of management_server since it seems to work either way."
This reverts commit a9496c8eee327bce87cc154f5ccde2f8ff65bf9b.
