+++ run.py
@@ -1,4 +1,4 @@
-from ufo import app
+from ufo.management_server import app
 
 if __name__ == '__main__':
   app.run(debug=True, host='0.0.0.0')
