+++ ufo/user.py
@@ -1,4 +1,4 @@
-from . import app, db, setup_required
+from management_server import app, db, setup_required
 
 import ast
 import base64
